Change: move create_alert and modify_alert out of manage_sql.c
