+++ src/manage_sql.c
@@ -6603,1307 +6603,6 @@ DEF_ACCESS (task_role_iterator_uuid, 4);
 /* Events and Alerts. */
 
 /**
- * @brief Validate an email address.
- *
- * @param[in]  address  Email address.
- *
- * @return 0 success, 1 failure.
- */
-static int
-validate_email (const char* address)
-{
-  gchar **split, *point;
-
-  assert (address);
-
-  split = g_strsplit (address, "@", 0);
-
-  if (split[0] == NULL || split[1] == NULL || split[2])
-    {
-      g_strfreev (split);
-      return 1;
-    }
-
-  /* Local part. */
-  point = split[0];
-  while (*point)
-    if (isalnum (*point)
-        || strchr ("!#$%&'*+-/=?^_`{|}~", *point)
-        || ((*point == '.')
-            && (point > split[0])
-            && point[1]
-            && (point[1] != '.')
-            && (point[-1] != '.')))
-      point++;
-    else
-      {
-        g_strfreev (split);
-        return 1;
-      }
-
-  /* Domain. */
-  point = split[1];
-  while (*point)
-    if (isalnum (*point)
-        || strchr ("-_", *point)  /* RFC actually forbids _. */
-        || ((*point == '.')
-            && (point > split[1])
-            && point[1]
-            && (point[1] != '.')
-            && (point[-1] != '.')))
-      point++;
-    else
-      {
-        g_strfreev (split);
-        return 1;
-      }
-
-  g_strfreev (split);
-  return 0;
-}
-
-/**
- * @brief Validate an email address list.
- *
- * @param[in]  list  Comma separated list of email addresses.
- *
- * @return 0 success, 1 failure.
- */
-static int
-validate_email_list (const char *list)
-{
-  gchar **split, **point;
-
-  assert (list);
-
-  split = g_strsplit (list, ",", 0);
-
-  if (split[0] == NULL)
-    {
-      g_strfreev (split);
-      return 1;
-    }
-
-  point = split;
-  while (*point)
-    {
-      const char *address;
-      address = *point;
-      while (*address && (*address == ' ')) address++;
-      if (validate_email (address))
-        {
-          g_strfreev (split);
-          return 1;
-        }
-      point++;
-    }
-
-  g_strfreev (split);
-  return 0;
-}
-
-/**
- * @brief Validate condition data for an alert.
- *
- * @param[in]  name      Name.
- * @param[in]  data      Data to validate.
- * @param[in]  condition The condition.
- *
- * @return 0 on success, 1 unexpected data name, 2 syntax error in data,
- *         3 failed to find filter for condition, -1 internal error.
- */
-static int
-validate_alert_condition_data (gchar *name, gchar* data,
-                               alert_condition_t condition)
-{
-  if (condition == ALERT_CONDITION_ALWAYS)
-    return 1;
-  if (condition == ALERT_CONDITION_SEVERITY_AT_LEAST)
-    {
-      if (strcmp (name, "severity"))
-        return 1;
-
-      if (g_regex_match_simple ("^(-1(\\.0)?|[0-9](\\.[0-9])?|10(\\.0))$",
-                                data ? data : "",
-                                0,
-                                0)
-          == 0)
-        return 2;
-    }
-  else if (condition == ALERT_CONDITION_SEVERITY_CHANGED)
-    {
-      if (strcmp (name, "direction"))
-        return 1;
-
-      if (g_regex_match_simple ("^(increased|decreased|changed)$",
-                                data ? data : "",
-                                0,
-                                0)
-          == 0)
-        return 2;
-    }
-  else if (condition == ALERT_CONDITION_FILTER_COUNT_AT_LEAST)
-    {
-      if (strcmp (name, "filter_id") == 0)
-        {
-          filter_t filter;
-          if (data == NULL)
-            return 3;
-          filter = 0;
-          if (find_filter_with_permission (data, &filter, "get_filters"))
-            return -1;
-          if (filter == 0)
-            return 3;
-          return 0;
-        }
-
-      if (strcmp (name, "count"))
-        return 1;
-    }
-  else if (condition == ALERT_CONDITION_FILTER_COUNT_CHANGED)
-    {
-      if (strcmp (name, "filter_id") == 0)
-        {
-          filter_t filter;
-          if (data == NULL)
-            return 3;
-          filter = 0;
-          if (find_filter_with_permission (data, &filter, "get_filters"))
-            return -1;
-          if (filter == 0)
-            return 3;
-          return 0;
-        }
-
-      if (strcmp (name, "direction")
-          && strcmp (name, "count"))
-        return 1;
-
-      if (strcmp (name, "direction") == 0
-          && g_regex_match_simple ("^(increased|decreased|changed)$",
-                                   data ? data : "",
-                                   0,
-                                   0)
-             == 0)
-        return 2;
-    }
-
-
-  return 0;
-}
-
-/**
- * @brief Validate event data for an alert.
- *
- * @param[in]  name   Name.
- * @param[in]  data   Data to validate.
- * @param[in]  event  The event.
- *
- * @return 0 on success, 1 unexpected data name, 2 syntax error in data.
- */
-static int
-validate_alert_event_data (gchar *name, gchar* data, event_t event)
-{
-  if (event == EVENT_NEW_SECINFO || event == EVENT_UPDATED_SECINFO)
-    {
-      if (strcmp (name, "secinfo_type"))
-        return 1;
-
-      if (data == NULL)
-        return 2;
-
-      if (strcasecmp (data, "nvt")
-          && strcasecmp (data, "cve")
-          && strcasecmp (data, "cpe")
-          && strcasecmp (data, "cert_bund_adv")
-          && strcasecmp (data, "dfn_cert_adv"))
-        return 2;
-    }
-  return 0;
-}
-
-/**
- * @brief Validate method data for the email method.
- *
- * @param[in]  method          Method that data corresponds to.
- * @param[in]  name            Name of data.
- * @param[in]  data            The data.
- * @param[in]  for_modify      Whether to return error codes for modify_alert.
- *
- * @return 0 valid, 2 or 6: validation of email address failed,
- *         7 or 9 subject too long, 8 or 10 message too long,
- *         60 recipient credential not found, 61 invalid recipient credential
- *         type, -1 error. When for_modify is 0, the first code is returned,
- *         otherwise the second one.
- */
-int
-validate_email_data (alert_method_t method, const gchar *name, gchar **data,
-                     int for_modify)
-{
-  if (method == ALERT_METHOD_EMAIL
-      && strcmp (name, "to_address") == 0
-      && validate_email_list (*data))
-    return for_modify ? 6 : 2;
-
-  if (method == ALERT_METHOD_EMAIL
-      && strcmp (name, "from_address") == 0
-      && validate_email (*data))
-    return for_modify ? 6 : 2;
-
-  if (method == ALERT_METHOD_EMAIL
-      && strcmp (name, "subject") == 0
-      && strlen (*data) > 80)
-    return for_modify ? 9 : 7;
-
-  if (method == ALERT_METHOD_EMAIL
-      && strcmp (name, "message") == 0
-      && strlen (*data) > get_max_email_message_size ())
-    return for_modify ? 10 : 8;
-
-  if (method == ALERT_METHOD_EMAIL
-      && strcmp (name, "recipient_credential") == 0
-      && *data && strcmp (*data, ""))
-    {
-      credential_t credential;
-      char *type;
-
-      if (find_credential_with_permission (*data, &credential, NULL))
-        return -1;
-      else if (credential == 0)
-        return 60;
-
-      type = credential_type (credential);
-      if (strcmp (type, "pgp")
-          && strcmp (type, "smime"))
-        {
-          free (type);
-          return 61;
-        }
-      free (type);
-    }
-
-  return 0;
-}
-
-/**
- * @brief Validate method data for the SCP method.
- *
- * @param[in]  method          Method that data corresponds to.
- * @param[in]  name            Name of data.
- * @param[in]  data            The data.
- *
- * @return 0 valid, 15 error in SCP host, 16 error in SCP port,
- *         17 failed to find report format for SCP method,
- *         18 error in SCP credential, 19 error in SCP path,
- *         -1 error.
- */
-static int
-validate_scp_data (alert_method_t method, const gchar *name, gchar **data)
-{
-  if (method == ALERT_METHOD_SCP
-      && strcmp (name, "scp_credential") == 0)
-    {
-      credential_t credential;
-      if (find_credential_with_permission (*data, &credential,
-                                           "get_credentials"))
-        return -1;
-      else if (credential == 0)
-        return 18;
-      else
-        {
-          gchar *username;
-          username = credential_value (credential, "username");
-
-          if (username == NULL || strlen (username) == 0)
-            {
-              g_free (username);
-              return 18;
-            }
-
-          if (strchr (username, ':'))
-            {
-              g_free (username);
-              return 18;
-            }
-
-          g_free (username);
-        }
-    }
-
-  if (method == ALERT_METHOD_SCP
-      && strcmp (name, "scp_path") == 0)
-    {
-      if (strlen (*data) == 0)
-        return 19;
-    }
-
-  if (method == ALERT_METHOD_SCP
-      && strcmp (name, "scp_host") == 0)
-    {
-      int type;
-      gchar *stripped;
-
-      stripped = g_strstrip (g_strdup (*data));
-      type = gvm_get_host_type (stripped);
-      g_free (stripped);
-      if ((type != HOST_TYPE_IPV4)
-          && (type != HOST_TYPE_IPV6)
-          && (type != HOST_TYPE_NAME))
-        return 15;
-    }
-
-  if (method == ALERT_METHOD_SCP
-      && strcmp (name, "scp_port") == 0)
-    {
-      int port;
-
-      port = atoi (*data);
-      if (port <= 0 || port > 65535)
-        return 16;
-    }
-
-  if (method == ALERT_METHOD_SCP
-      && strcmp (name, "scp_report_format") == 0)
-    {
-      report_format_t report_format;
-
-      report_format = 0;
-      if (find_report_format_with_permission (*data,
-                                              &report_format,
-                                              "get_report_formats"))
-        return -1;
-      if (report_format == 0)
-        return 17;
-    }
-
-  return 0;
-}
-
-/**
- * @brief Validate method data for the Send method.
- *
- * @param[in]  method          Method that data corresponds to.
- * @param[in]  name            Name of data.
- * @param[in]  data            The data.
- *
- * @return 0 valid, 12 error in Send host, 13 error in Send port, 14 failed
- *         to find report format for Send method, -1 error.
- */
-static int
-validate_send_data (alert_method_t method, const gchar *name, gchar **data)
-{
-  if (method == ALERT_METHOD_SEND
-      && strcmp (name, "send_host") == 0)
-    {
-      int type;
-      gchar *stripped;
-
-      stripped = g_strstrip (g_strdup (*data));
-      type = gvm_get_host_type (stripped);
-      g_free (stripped);
-      if ((type != HOST_TYPE_IPV4)
-          && (type != HOST_TYPE_IPV6)
-          && (type != HOST_TYPE_NAME))
-        return 12;
-    }
-
-  if (method == ALERT_METHOD_SEND
-      && strcmp (name, "send_port") == 0)
-    {
-      int port;
-      gchar *stripped, *end;
-
-      stripped = g_strstrip (g_strdup (*data));
-      port = strtol (stripped, &end, 10);
-      if (*end != '\0')
-        {
-          g_free (stripped);
-          return 13;
-        }
-
-      g_free (stripped);
-      g_free (*data);
-      *data = g_strdup_printf ("%i", port);
-    }
-
-  if (method == ALERT_METHOD_SEND
-      && strcmp (name, "send_report_format") == 0)
-    {
-      report_format_t report_format;
-
-      report_format = 0;
-      if (find_report_format_with_permission (*data,
-                                              &report_format,
-                                              "get_report_formats"))
-        return -1;
-      if (report_format == 0)
-        return 14;
-    }
-
-  return 0;
-}
-
-/**
- * @brief Validate method data for the Send method.
- *
- * @param[in]  method          Method that data corresponds to.
- * @param[in]  name            Name of data.
- * @param[in]  data            The data.
- *
- * @return 0 valid, 40 invalid credential, 41 invalid SMB share path,
- *         42 invalid SMB file path, 43 SMB file path contains dot, -1 error.
- */
-static int
-validate_smb_data (alert_method_t method, const gchar *name, gchar **data)
-{
-  if (method == ALERT_METHOD_SMB)
-    {
-      if (strcmp (name, "smb_credential") == 0)
-        {
-          credential_t credential;
-          if (find_credential_with_permission (*data, &credential,
-                                              "get_credentials"))
-            return -1;
-          else if (credential == 0)
-            return 40;
-          else
-            {
-              gchar *username;
-              username = credential_value (credential, "username");
-
-              if (username == NULL || strlen (username) == 0)
-                {
-                  g_free (username);
-                  return 40;
-                }
-
-              if (strchr (username, '@') || strchr (username, ':'))
-                {
-                  g_free (username);
-                  return 40;
-                }
-
-              g_free (username);
-            }
-        }
-
-      if (strcmp (name, "smb_share_path") == 0)
-        {
-          /* Check if share path has the correct format
-           *  "\\<host>\<share>" */
-          if (g_regex_match_simple ("^(?>\\\\\\\\|\\/\\/)[^:?<>|]+"
-                                    "(?>\\\\|\\/)[^:?<>|]+$", *data, 0, 0)
-              == FALSE)
-            {
-              return 41;
-            }
-        }
-
-      if (strcmp (name, "smb_file_path") == 0)
-        {
-          /* Check if file path contains invalid characters:
-           *  ":", "?", "<", ">", "|" */
-          if (g_regex_match_simple ("^[^:?<>|]+$", *data, 0, 0)
-              == FALSE)
-            {
-              return 42;
-            }
-          /* Check if a file or directory name ends with a dot,
-           *  e.g. "../a", "abc/../xyz" or "abc/..". */
-          else if (g_regex_match_simple ("^(?:.*\\.)(?:[\\/\\\\].*)*$",
-                                         *data, 0, 0))
-            {
-              return 43;
-            }
-        }
-
-    }
-
-  return 0;
-}
-
-/**
- * @brief Validate method data for the TippingPoint method.
- *
- * @param[in]  method          Method that data corresponds to.
- * @param[in]  name            Name of data.
- * @param[in]  data            The data.
- *
- * @return 0 valid, 50 invalid credential, 51 invalid hostname,
- *  52 invalid certificate, 53 invalid TLS workaround setting.
- */
-static int
-validate_tippingpoint_data (alert_method_t method, const gchar *name,
-                             gchar **data)
-{
-  if (method == ALERT_METHOD_TIPPINGPOINT)
-    {
-      if (strcmp (name, "tp_sms_credential") == 0)
-        {
-          credential_t credential;
-          if (find_credential_with_permission (*data, &credential,
-                                               "get_credentials"))
-            return -1;
-          else if (credential == 0)
-            return 50;
-          else
-            {
-              if (strcmp (credential_type (credential), "up"))
-                return 50;
-
-            }
-        }
-
-      if (strcmp (name, "tp_sms_hostname") == 0)
-        {
-          if (g_regex_match_simple ("^[0-9A-Za-z][0-9A-Za-z.\\-]*$",
-                                    *data, 0, 0)
-              == FALSE)
-            {
-              return 51;
-            }
-        }
-
-      if (strcmp (name, "tp_sms_tls_certificate") == 0)
-        {
-          // Check certificate, return 52 on failure
-          int ret;
-          gnutls_x509_crt_fmt_t crt_fmt;
-
-          ret = get_certificate_info (*data, strlen(*data), FALSE,
-                                      NULL, NULL, NULL,
-                                      NULL, NULL, NULL, NULL, &crt_fmt);
-          if (ret || crt_fmt != GNUTLS_X509_FMT_PEM)
-            {
-              return 52;
-            }
-        }
-
-      if (strcmp (name, "tp_sms_tls_workaround") == 0)
-        {
-          if (g_regex_match_simple ("^0|1$", *data, 0, 0)
-              == FALSE)
-            {
-              return 53;
-            }
-        }
-    }
-
-  return 0;
-}
-
-/**
- * @brief Validate method data for the vFire alert method.
- *
- * @param[in]  method          Method that data corresponds to.
- * @param[in]  name            Name of data.
- * @param[in]  data            The data.
- *
- * @return 0 valid, 70 credential not found, 71 invalid credential type
- */
-static int
-validate_vfire_data (alert_method_t method, const gchar *name,
-                     gchar **data)
-{
-  if (method == ALERT_METHOD_VFIRE)
-    {
-      if (strcmp (name, "vfire_credential") == 0)
-        {
-          credential_t credential;
-          if (find_credential_with_permission (*data, &credential,
-                                               "get_credentials"))
-            return -1;
-          else if (credential == 0)
-            return 70;
-          else
-            {
-              char *cred_type = credential_type (credential);
-              if (strcmp (cred_type, "up"))
-                {
-                  free (cred_type);
-                  return 71;
-                }
-              free (cred_type);
-            }
-        }
-    }
-  return 0;
-}
-
-/**
- * @brief Validate method data for the Sourcefire method.
- *
- * @param[in]  method          Method that data corresponds to.
- * @param[in]  name            Name of data.
- * @param[in]  data            The data.
- *
- * @return 0 valid, 80 credential not found, 81 invalid credential type
- */
-static int
-validate_sourcefire_data (alert_method_t method, const gchar *name,
-                          gchar **data)
-{
-  if (method == ALERT_METHOD_SOURCEFIRE)
-    {
-      if (strcmp (name, "pkcs12_credential") == 0)
-        {
-          credential_t credential;
-          if (find_credential_with_permission (*data, &credential,
-                                               "get_credentials"))
-            return -1;
-          else if (credential == 0)
-            return 80;
-          else
-            {
-              char *sourcefire_credential_type;
-              sourcefire_credential_type = credential_type (credential);
-              if (strcmp (sourcefire_credential_type, "up")
-                  && strcmp (sourcefire_credential_type, "pw"))
-                {
-                  free (sourcefire_credential_type);
-                  return 81;
-                }
-              free (sourcefire_credential_type);
-            }
-        }
-    }
-
-  return 0;
-}
-
-/**
- * @brief Check alert params.
- *
- * @param[in]  event           Type of event.
- * @param[in]  condition       Event condition.
- * @param[in]  method          Escalation method.
- *
- * @return 0 success, 20 method does not match event, 21 condition does not
- *         match event.
- */
-static int
-check_alert_params (event_t event, alert_condition_t condition,
-                    alert_method_t method)
-{
-  if (event == EVENT_NEW_SECINFO || event == EVENT_UPDATED_SECINFO)
-    {
-      if (method == ALERT_METHOD_HTTP_GET
-          || method == ALERT_METHOD_SOURCEFIRE
-          || method == ALERT_METHOD_VERINICE)
-        return 20;
-
-      if (condition == ALERT_CONDITION_SEVERITY_AT_LEAST
-          || condition == ALERT_CONDITION_SEVERITY_CHANGED
-          || condition == ALERT_CONDITION_FILTER_COUNT_CHANGED)
-        return 21;
-    }
-  return 0;
-}
-
-/**
- * @brief Create an alert.
- *
- * @param[in]  name            Name of alert.
- * @param[in]  comment         Comment on alert.
- * @param[in]  filter_id       Filter.
- * @param[in]  active          Whether the alert is active.
- * @param[in]  event           Type of event.
- * @param[in]  event_data      Type-specific event data.
- * @param[in]  condition       Event condition.
- * @param[in]  condition_data  Condition-specific data.
- * @param[in]  method          Escalation method.
- * @param[in]  method_data     Data for escalation method.
- * @param[out] alert       Created alert on success.
- *
- * @return 0 success, 1 escalation exists already, 2 validation of email failed,
- *         3 failed to find filter, 4 type must be "result" if specified,
- *         5 unexpected condition data name, 6 syntax error in condition data,
- *         7 email subject too long, 8 email message too long, 9 failed to find
- *         filter for condition, 12 error in Send host, 13 error in Send port,
- *         14 failed to find report format for Send method,
- *         15 error in SCP host, 16 error in SCP port,
- *         17 failed to find report format for SCP method, 18 error
- *         in SCP credential, 19 error in SCP path, 20 method does not match
- *         event, 21 condition does not match event, 31 unexpected event data
- *         name, 32 syntax error in event data, 40 invalid SMB credential
- *       , 41 invalid SMB share path, 42 invalid SMB file path,
- *         43 SMB file path contains dot,
- *         50 invalid TippingPoint credential, 51 invalid TippingPoint hostname,
- *         52 invalid TippingPoint certificate, 53 invalid TippingPoint TLS
- *         workaround setting, 60 recipient credential not found, 61 invalid
- *         recipient credential type, 70 vFire credential not found,
- *         71 invalid vFire credential type,
- *         99 permission denied, -1 error.
- */
-int
-create_alert (const char* name, const char* comment, const char* filter_id,
-              const char* active, event_t event, GPtrArray* event_data,
-              alert_condition_t condition, GPtrArray* condition_data,
-              alert_method_t method, GPtrArray* method_data,
-              alert_t *alert)
-{
-  int index, ret;
-  gchar *item, *quoted_comment;
-  gchar *quoted_name;
-  filter_t filter;
-
-  assert (current_credentials.uuid);
-
-  sql_begin_immediate ();
-
-  if (acl_user_may ("create_alert") == 0)
-    {
-      sql_rollback ();
-      return 99;
-    }
-
-  ret = check_alert_params (event, condition, method);
-  if (ret)
-    {
-      sql_rollback ();
-      return ret;
-    }
-
-  filter = 0;
-  if (event != EVENT_NEW_SECINFO && event != EVENT_UPDATED_SECINFO && filter_id
-      && strcmp (filter_id, "0"))
-    {
-      char *type;
-
-      if (find_filter_with_permission (filter_id, &filter, "get_filters"))
-        {
-          sql_rollback ();
-          return -1;
-        }
-
-      if (filter == 0)
-        {
-          sql_rollback ();
-          return 3;
-        }
-
-      /* Filter type must be result if specified. */
-
-      type = sql_string ("SELECT type FROM filters WHERE id = %llu;",
-                         filter);
-      if (type && strcasecmp (type, "result"))
-        {
-          free (type);
-          sql_rollback ();
-          return 4;
-        }
-      free (type);
-    }
-
-  if (resource_with_name_exists (name, "alert", 0))
-    {
-      sql_rollback ();
-      return 1;
-    }
-  quoted_name = sql_quote (name);
-  quoted_comment = sql_quote (comment ?: "");
-
-  sql ("INSERT INTO alerts (uuid, owner, name, comment, event, condition,"
-       " method, filter, active, creation_time, modification_time)"
-       " VALUES (make_uuid (),"
-       " (SELECT id FROM users WHERE users.uuid = '%s'),"
-       " '%s', '%s', %i, %i, %i, %llu, %i, m_now (), m_now ());",
-       current_credentials.uuid,
-       quoted_name,
-       quoted_comment,
-       event,
-       condition,
-       method,
-       filter,
-       active ? strcmp (active, "0") : 1);
-
-  g_free (quoted_comment);
-  g_free (quoted_name);
-
-  *alert = sql_last_insert_id ();
-
-  index = 0;
-  while ((item = (gchar*) g_ptr_array_index (condition_data, index++)))
-    {
-      int validation_result;
-      gchar *data_name = sql_quote (item);
-      gchar *data = sql_quote (item + strlen (item) + 1);
-
-      validation_result = validate_alert_condition_data (data_name,
-                                                         data,
-                                                         condition);
-
-      if (validation_result)
-        {
-          g_free (data_name);
-          g_free (data);
-          sql_rollback ();
-
-          switch (validation_result)
-            {
-              case 1:
-                return 5;
-              case 2:
-                return 6;
-              case 3:
-                return 9;
-              default:
-                return -1;
-            }
-        }
-
-      sql ("INSERT INTO alert_condition_data (alert, name, data)"
-           " VALUES (%llu, '%s', '%s');",
-           *alert,
-           data_name,
-           data);
-      g_free (data_name);
-      g_free (data);
-    }
-
-  index = 0;
-  while ((item = (gchar*) g_ptr_array_index (event_data, index++)))
-    {
-      int validation_result;
-      gchar *data_name = sql_quote (item);
-      gchar *data = sql_quote (item + strlen (item) + 1);
-
-      validation_result = validate_alert_event_data (data_name, data, event);
-
-      if (validation_result)
-        {
-          g_free (data_name);
-          g_free (data);
-          sql_rollback ();
-
-          switch (validation_result)
-            {
-              case 1:
-                return 31;
-              case 2:
-                return 32;
-              default:
-                return -1;
-            }
-        }
-
-      sql ("INSERT INTO alert_event_data (alert, name, data)"
-           " VALUES (%llu, '%s', '%s');",
-           *alert,
-           data_name,
-           data);
-      g_free (data_name);
-      g_free (data);
-    }
-
-  index = 0;
-  while ((item = (gchar*) g_ptr_array_index (method_data, index++)))
-    {
-      gchar *data_name, *data;
-
-      data_name = sql_quote (item);
-      data = sql_quote (item + strlen (item) + 1);
-
-      ret = validate_email_data (method, data_name, &data, 0);
-      if (ret)
-        {
-          g_free (data_name);
-          g_free (data);
-          sql_rollback ();
-          return ret;
-        }
-
-      ret = validate_scp_data (method, data_name, &data);
-      if (ret)
-        {
-          g_free (data_name);
-          g_free (data);
-          sql_rollback ();
-          return ret;
-        }
-
-      ret = validate_send_data (method, data_name, &data);
-      if (ret)
-        {
-          g_free (data_name);
-          g_free (data);
-          sql_rollback ();
-          return ret;
-        }
-
-      ret = validate_smb_data (method, data_name, &data);
-      if (ret)
-        {
-          g_free (data_name);
-          g_free (data);
-          sql_rollback ();
-          return ret;
-        }
-
-      ret = validate_sourcefire_data (method, data_name, &data);
-      if (ret)
-        {
-          g_free (data_name);
-          g_free (data);
-          sql_rollback ();
-          return ret;
-        }
-
-      ret = validate_tippingpoint_data (method, data_name, &data);
-      if (ret)
-        {
-          g_free (data_name);
-          g_free (data);
-          sql_rollback ();
-          return ret;
-        }
-
-      ret = validate_vfire_data (method, data_name, &data);
-      if (ret)
-        {
-          g_free (data_name);
-          g_free (data);
-          sql_rollback ();
-          return ret;
-        }
-
-      sql ("INSERT INTO alert_method_data (alert, name, data)"
-           " VALUES (%llu, '%s', '%s');",
-           *alert,
-           data_name,
-           data);
-      g_free (data_name);
-      g_free (data);
-    }
-
-  sql_commit ();
-
-  return 0;
-}
-
-/**
- * @brief Modify an alert.
- *
- * @param[in]   alert_id        UUID of alert.
- * @param[in]   name            Name of alert.
- * @param[in]   comment         Comment on alert.
- * @param[in]   filter_id       Filter.
- * @param[in]   active          Whether the alert is active.  NULL to leave it
- *                              at the current value.
- * @param[in]   event           Type of event.
- * @param[in]   event_data      Type-specific event data.
- * @param[in]   condition       Event condition.
- * @param[in]   condition_data  Condition-specific data.
- * @param[in]   method          Escalation method.
- * @param[in]   method_data     Data for escalation method.
- *
- * @return 0 success, 1 failed to find alert, 2 alert with new name exists,
- *         3 alert_id required, 4 failed to find filter, 5 filter type must be
- *         result if specified, 6 Provided email address not valid,
- *         7 unexpected condition data name, 8 syntax error in condition data,
- *         9 email subject too long, 10 email message too long, 11 failed to
- *         find filter for condition, 12 error in Send host, 13 error in Send
- *         port, 14 failed to find report format for Send method,
- *         15 error in SCP host, 16 error in SCP port,
- *         17 failed to find report format for SCP method, 18 error
- *         in SCP credential, 19 error in SCP path, 20 method does not match
- *         event, 21 condition does not match event, 31 unexpected event data
- *         name, 32 syntax error in event data, 40 invalid SMB credential
- *       , 41 invalid SMB share path, 42 invalid SMB file path,
- *         43 SMB file path contains dot,
- *         50 invalid TippingPoint credential, 51 invalid TippingPoint hostname,
- *         52 invalid TippingPoint certificate, 53 invalid TippingPoint TLS
- *         workaround setting, 60 recipient credential not found, 61 invalid
- *         recipient credential type, 70 vFire credential not found,
- *         71 invalid vFire credential type,
- *         99 permission denied, -1 internal error.
- */
-int
-modify_alert (const char *alert_id, const char *name, const char *comment,
-              const char *filter_id, const char *active, event_t event,
-              GPtrArray *event_data, alert_condition_t condition,
-              GPtrArray *condition_data, alert_method_t method,
-              GPtrArray *method_data)
-{
-  int index, ret;
-  gchar *quoted_name, *quoted_comment, *item;
-  alert_t alert;
-  filter_t filter;
-
-  if (alert_id == NULL)
-    return 3;
-
-  sql_begin_immediate ();
-
-  assert (current_credentials.uuid);
-
-  if (acl_user_may ("modify_alert") == 0)
-    {
-      sql_rollback ();
-      return 99;
-    }
-
-  ret = check_alert_params (event, condition, method);
-  if (ret)
-    {
-      sql_rollback ();
-      return ret;
-    }
-
-  alert = 0;
-  if (find_alert_with_permission (alert_id, &alert, "modify_alert"))
-    {
-      sql_rollback ();
-      return -1;
-    }
-
-  if (alert == 0)
-    {
-      sql_rollback ();
-      return 1;
-    }
-
-  /* Check whether an alert with the same name exists already. */
-  if (resource_with_name_exists (name, "alert", alert))
-    {
-      sql_rollback ();
-      return 2;
-    }
-
-  /* Check filter. */
-  filter = 0;
-  if (event != EVENT_NEW_SECINFO && event != EVENT_UPDATED_SECINFO && filter_id
-      && strcmp (filter_id, "0"))
-    {
-      char *type;
-
-      if (find_filter_with_permission (filter_id, &filter, "get_filters"))
-        {
-          sql_rollback ();
-          return -1;
-        }
-
-      if (filter == 0)
-        {
-          sql_rollback ();
-          return 4;
-        }
-
-      /* Filter type must be report if specified. */
-
-      type = sql_string ("SELECT type FROM filters WHERE id = %llu;",
-                         filter);
-      if (type && strcasecmp (type, "result"))
-        {
-          free (type);
-          sql_rollback ();
-          return 5;
-        }
-      free (type);
-    }
-
-  quoted_name = sql_quote (name ?: "");
-  quoted_comment = sql_quote (comment ? comment : "");
-
-  sql ("UPDATE alerts SET"
-       " name = '%s',"
-       " comment = '%s',"
-       " filter = %llu,"
-       " active = %s,"
-       " modification_time = m_now ()"
-       " WHERE id = %llu;",
-       quoted_name,
-       quoted_comment,
-       filter,
-       active
-        ? (strcmp (active, "0") ? "1" : "0")
-        : "active",
-       alert);
-
-  g_free (quoted_comment);
-  g_free (quoted_name);
-
-  /* Modify alert event */
-  if (event != EVENT_ERROR)
-    {
-      sql ("UPDATE alerts set event = %i WHERE id = %llu", event, alert);
-      sql ("DELETE FROM alert_event_data WHERE alert = %llu", alert);
-      index = 0;
-      while ((item = (gchar*) g_ptr_array_index (event_data, index++)))
-        {
-          int validation_result;
-          gchar *data_name = sql_quote (item);
-          gchar *data = sql_quote (item + strlen (item) + 1);
-
-          validation_result = validate_alert_event_data (data_name,
-                                                         data,
-                                                         event);
-
-          if (validation_result)
-            {
-              g_free (data_name);
-              g_free (data);
-              sql_rollback ();
-
-              switch (validation_result)
-                {
-                  case 1:
-                    return 31;
-                  case 2:
-                    return 32;
-                  default:
-                    return -1;
-                }
-            }
-
-          sql ("INSERT INTO alert_event_data (alert, name, data)"
-               " VALUES (%llu, '%s', '%s');",
-               alert,
-               data_name,
-               data);
-          g_free (data_name);
-          g_free (data);
-        }
-    }
-
-  /* Modify alert condition */
-  if (condition != ALERT_CONDITION_ERROR)
-    {
-      sql ("UPDATE alerts set condition = %i WHERE id = %llu",
-           condition,
-           alert);
-      sql ("DELETE FROM alert_condition_data WHERE alert = %llu", alert);
-      index = 0;
-      while ((item = (gchar*) g_ptr_array_index (condition_data, index++)))
-        {
-          int validation_result;
-          gchar *data_name = sql_quote (item);
-          gchar *data = sql_quote (item + strlen (item) + 1);
-
-          validation_result = validate_alert_condition_data (data_name, data,
-                                                             condition);
-
-          if (validation_result)
-            {
-              g_free (data_name);
-              g_free (data);
-              sql_rollback ();
-
-              switch (validation_result)
-                {
-                  case 1:
-                    return 7;
-                  case 2:
-                    return 8;
-                  case 3:
-                    return 11;
-                  default:
-                    return -1;
-                }
-            }
-
-          sql ("INSERT INTO alert_condition_data (alert, name, data)"
-               " VALUES (%llu, '%s', '%s');",
-               alert,
-               data_name,
-               data);
-          g_free (data_name);
-          g_free (data);
-        }
-    }
-
-  /* Modify alert method */
-  if (method != ALERT_METHOD_ERROR)
-    {
-      sql ("UPDATE alerts set method = %i WHERE id = %llu", method, alert);
-      sql ("DELETE FROM alert_method_data WHERE alert = %llu", alert);
-      index = 0;
-      while ((item = (gchar*) g_ptr_array_index (method_data, index++)))
-        {
-          gchar *data_name, *data;
-
-          data_name = sql_quote (item);
-          data = sql_quote (item + strlen (item) + 1);
-
-          ret = validate_email_data (method, data_name, &data, 1);
-          if (ret)
-            {
-              g_free (data_name);
-              g_free (data);
-              sql_rollback ();
-              return ret;
-            }
-
-          ret = validate_scp_data (method, data_name, &data);
-          if (ret)
-            {
-              g_free (data_name);
-              g_free (data);
-              sql_rollback ();
-              return ret;
-            }
-
-          ret = validate_send_data (method, data_name, &data);
-          if (ret)
-            {
-              g_free (data_name);
-              g_free (data);
-              sql_rollback ();
-              return ret;
-            }
-
-          ret = validate_smb_data (method, data_name, &data);
-          if (ret)
-            {
-              g_free (data_name);
-              g_free (data);
-              sql_rollback ();
-              return ret;
-            }
-
-          ret = validate_sourcefire_data (method, data_name, &data);
-          if (ret)
-            {
-              g_free (data_name);
-              g_free (data);
-              sql_rollback ();
-              return ret;
-            }
-
-          ret = validate_tippingpoint_data (method, data_name, &data);
-          if (ret)
-            {
-              g_free (data_name);
-              g_free (data);
-              sql_rollback ();
-              return ret;
-            }
-
-          ret = validate_vfire_data (method, data_name, &data);
-          if (ret)
-            {
-              g_free (data_name);
-              g_free (data);
-              sql_rollback ();
-              return ret;
-            }
-
-          sql ("INSERT INTO alert_method_data (alert, name, data)"
-               " VALUES (%llu, '%s', '%s');",
-               alert,
-               data_name,
-               data);
-          g_free (data_name);
-          g_free (data);
-        }
-    }
-
-  sql_commit ();
-
-  return 0;
-}
-
-/**
  * @brief Return the name of an alert.
  *
  * @param[in]  alert  Alert.
